requirements.txt
google-generativeai
langchain
jupyter
dotenv
langchain-google-genai
pytubefix
langchain_community
scikit-learn
docarray
langchain_pinecone
tiktoken
botbuilder-core
aiohttp
fastapi
uvicorn
git+https://github.com/openai/whisper.git